requirements.txt
neat-python==0.92
pygame==2.1.2
matplotlib==3.5.2
